move core feel machinery to the sdk
