+++ connector-runtime/connector-runtime-core/pom.xml
@@ -16,7 +16,6 @@
 
   <properties>
     <system-stubs.version>2.0.2</system-stubs.version>
-    <jsonassert.version>1.5.1</jsonassert.version>
   </properties>
 
   <dependencies>
@@ -73,13 +72,6 @@
     <dependency>
       <groupId>org.assertj</groupId>
       <artifactId>assertj-core</artifactId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.skyscreamer</groupId>
-      <artifactId>jsonassert</artifactId>
-      <version>${jsonassert.version}</version>
       <scope>test</scope>
     </dependency>
