+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/ConnectorHelper.java
@@ -21,8 +21,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.connector.api.error.BpmnError;
 import io.camunda.connector.api.inbound.InboundConnectorExecutable;
 import io.camunda.connector.api.outbound.OutboundConnectorFunction;
-import io.camunda.connector.runtime.core.feel.FeelEngineWrapper;
-import io.camunda.connector.runtime.core.feel.FeelEngineWrapperException;
+import io.camunda.connector.impl.feel.FeelEngineWrapper;
+import io.camunda.connector.impl.feel.FeelEngineWrapperException;
 import java.lang.reflect.InvocationTargetException;
 import java.util.HashMap;
 import java.util.Map;
