+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/inbound/correlation/InboundCorrelationHandler.java
@@ -19,6 +19,8 @@ package io.camunda.connector.runtime.core.inbound.correlation;
 import io.camunda.connector.api.error.ConnectorException;
 import io.camunda.connector.api.inbound.InboundConnectorResult;
 import io.camunda.connector.impl.ConnectorInputException;
+import io.camunda.connector.impl.feel.FeelEngineWrapper;
+import io.camunda.connector.impl.feel.FeelEngineWrapperException;
 import io.camunda.connector.impl.inbound.MessageCorrelationPoint;
 import io.camunda.connector.impl.inbound.StartEventCorrelationPoint;
 import io.camunda.connector.impl.inbound.result.CorrelatedMessage;
@@ -28,8 +30,6 @@ import io.camunda.connector.impl.inbound.result.MessageCorrelationResult;
 import io.camunda.connector.impl.inbound.result.ProcessInstance;
 import io.camunda.connector.impl.inbound.result.StartEventCorrelationResult;
 import io.camunda.connector.runtime.core.ConnectorHelper;
-import io.camunda.connector.runtime.core.feel.FeelEngineWrapper;
-import io.camunda.connector.runtime.core.feel.FeelEngineWrapperException;
 import io.camunda.connector.runtime.core.inbound.InboundConnectorDefinitionImpl;
 import io.camunda.zeebe.client.ZeebeClient;
 import io.camunda.zeebe.client.api.response.ProcessInstanceEvent;
