--- connector-runtime/connector-runtime-core/src/test/java/io/camunda/connector/runtime/core/feel/jackson/FeelDeserializerTest.java
@@ -1,48 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information regarding copyright
- * ownership. Camunda licenses this file to you under the Apache License,
- * Version 2.0; you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.connector.runtime.core.feel.jackson;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import java.util.Map;
-import org.junit.jupiter.api.Test;
-
-public class FeelDeserializerTest {
-
-  private final ObjectMapper mapper = new ObjectMapper();
-
-  @Test
-  void feelDeserializer_deserialize() throws JsonProcessingException {
-    // given
-    String json = """
-        { "props": "= { result: \\"foobar\\" }" }
-        """;
-
-    // when
-    TargetTypeObject targetType = mapper.readValue(json, TargetTypeObject.class);
-
-    // then
-    assertThat(targetType.props.size()).isEqualTo(1);
-    assertThat(targetType.props.get("result")).isEqualTo("foobar");
-  }
-
-  private record TargetTypeObject(
-      @JsonDeserialize(using = FeelDeserializer.class) Map<String, String> props) {}
-}
