+++ connectors/webhook/src/main/java/io/camunda/connector/inbound/authorization/JWTChecker.java
@@ -19,8 +19,8 @@ import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.camunda.connector.impl.feel.FeelEngineWrapperException;
 import io.camunda.connector.inbound.model.JWTProperties;
-import io.camunda.connector.runtime.core.feel.FeelEngineWrapperException;
 import java.security.interfaces.ECPublicKey;
 import java.security.interfaces.RSAPublicKey;
 import java.util.ArrayList;
