+++ connectors/webhook/src/test/java/io/camunda/connector/inbound/authorization/JWTCheckerTest.java
@@ -12,9 +12,9 @@ import static org.junit.Assert.assertTrue;
 import com.auth0.jwk.Jwk;
 import com.auth0.jwk.JwkProvider;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.camunda.connector.impl.feel.FeelEngineWrapper;
 import io.camunda.connector.inbound.model.JWTProperties;
 import io.camunda.connector.inbound.utils.ObjectMapperSupplier;
-import io.camunda.connector.runtime.core.feel.FeelEngineWrapper;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
