+++ pom.xml
@@ -6,7 +6,7 @@
   <parent>
     <groupId>io.camunda.connector</groupId>
     <artifactId>connector-parent</artifactId>
-    <version>0.11.2</version>
+    <version>0.12.0-SNAPSHOT</version>
     <relativePath/>
   </parent>
